Some miscellaneous code cleanup
